move out kubernetes types
