+++ src/azure/getAzureMetadata.ts
@@ -2,7 +2,7 @@ import { window } from 'vscode';
 import safesh from 'shell-escape-tag';
 import { QuickPickItem } from 'vscode';
 import { kubernetesTools } from '../kubernetes/kubernetesTools';
-import { ClusterProvider, ConfigMap } from '../kubernetes/types/kubernetesTypes';
+import { ClusterProvider, ConfigMap } from '../types/kubernetes/kubernetesTypes';
 import { shell, ShellResult } from '../shell';
 import { parseJson } from '../utils/jsonUtils';
 import { AzureClusterProvider, AzureConstants } from './azureTools';
