+++ src/commands/fluxReconcileSource.ts
@@ -1,7 +1,7 @@
 import { window } from 'vscode';
 import { fluxTools } from '../flux/fluxTools';
-import { FluxSource } from '../flux/fluxCliTypes';
-import { KubernetesObjectKinds } from '../kubernetes/types/kubernetesTypes';
+import { FluxSource } from '../types/fluxCliTypes';
+import { KubernetesObjectKinds } from '../types/kubernetes/kubernetesTypes';
 import { BucketNode } from '../treeviews/nodes/bucketNode';
 import { GitRepositoryNode } from '../treeviews/nodes/gitRepositoryNode';
 import { OCIRepositoryNode } from '../treeviews/nodes/ociRepositoryNode';
