+++ src/commands/fluxReconcileWorkload.ts
@@ -1,7 +1,7 @@
 import { window } from 'vscode';
 import { fluxTools } from '../flux/fluxTools';
-import { FluxWorkload } from '../flux/fluxCliTypes';
-import { KubernetesObjectKinds } from '../kubernetes/types/kubernetesTypes';
+import { FluxWorkload } from '../types/fluxCliTypes';
+import { KubernetesObjectKinds } from '../types/kubernetes/kubernetesTypes';
 import { HelmReleaseNode } from '../treeviews/nodes/helmReleaseNode';
 import { KustomizationNode } from '../treeviews/nodes/kustomizationNode';
 import { refreshWorkloadsTreeView } from '../treeviews/treeViews';
