+++ src/git/gitInfo.ts
@@ -5,7 +5,7 @@ import { workspace, window } from 'vscode';
 import { makeSSHUrlFromGitUrl } from '../commands/createSource';
 import { checkGitVersion } from '../install';
 import { kubernetesTools } from '../kubernetes/kubernetesTools';
-import { GitRepository } from '../kubernetes/types/flux/gitRepository';
+import { GitRepository } from '../types/flux/gitRepository';
 import { shell } from '../shell';
