+++ src/globalState.ts
@@ -1,5 +1,5 @@
 import { ExtensionContext, window, workspace } from 'vscode';
-import { KnownClusterProviders } from './kubernetes/types/kubernetesTypes';
+import { KnownClusterProviders } from './types/kubernetes/kubernetesTypes';
 
 export interface ClusterMetadata {
 	azureResourceGroup?: string;
