+++ src/treeviews/dataProviders/templateDataProvider.ts
@@ -1,6 +1,6 @@
 import { kubernetesTools } from '../../kubernetes/kubernetesTools';
 import { sortByMetadataName } from '../../kubernetes/kubernetesUtils';
-import { GitOpsTemplate } from '../../kubernetes/types/flux/gitOpsTemplate';
+import { GitOpsTemplate } from '../../types/flux/gitOpsTemplate';
 import { GitOpsTemplateNode } from '../nodes/gitOpsTemplateNode';
 import { DataProvider } from './dataProvider';
