+++ src/treeviews/nodes/clusterContextNode.ts
@@ -3,9 +3,9 @@ import { CommandId } from '../../commands/commands';
 import { globalState } from '../../extension';
 import { getExtensionContext } from '../../extensionContext';
 import { extensionState } from '../../extensionState';
-import { KubernetesCluster, KubernetesContextWithCluster } from '../../kubernetes/types/kubernetesConfig';
+import { KubernetesCluster, KubernetesContextWithCluster } from '../../types/kubernetes/kubernetesConfig';
 import { kubernetesTools } from '../../kubernetes/kubernetesTools';
-import { ClusterProvider } from '../../kubernetes/types/kubernetesTypes';
+import { ClusterProvider } from '../../types/kubernetes/kubernetesTypes';
 import { createMarkdownHr, createMarkdownTable } from '../../utils/markdownUtils';
 import { ContextTypes, setVSCodeContext } from '../../vscodeContext';
 import { NodeContext } from './nodeContext';
