+++ src/treeviews/nodes/gitRepositoryNode.ts
@@ -1,5 +1,5 @@
-import { GitRepository } from '../../kubernetes/types/flux/gitRepository';
-import { KubernetesObjectKinds } from '../../kubernetes/types/kubernetesTypes';
+import { GitRepository } from '../../types/flux/gitRepository';
+import { KubernetesObjectKinds } from '../../types/kubernetes/kubernetesTypes';
 import { NodeContext } from './nodeContext';
 import { SourceNode } from './sourceNode';
