+++ src/treeviews/nodes/helmRepositoryNode.ts
@@ -1,5 +1,5 @@
-import { HelmRepository } from '../../kubernetes/types/flux/helmRepository';
-import { KubernetesObjectKinds } from '../../kubernetes/types/kubernetesTypes';
+import { HelmRepository } from '../../types/flux/helmRepository';
+import { KubernetesObjectKinds } from '../../types/kubernetes/kubernetesTypes';
 import { NodeContext } from './nodeContext';
 import { SourceNode } from './sourceNode';
