+++ src/types/flux/helmRelease.ts
@@ -1,4 +1,4 @@
-import { DeploymentCondition, KubernetesJSON, KubernetesObject, KubernetesObjectKinds, ObjectMeta, ResultMetadata } from '../kubernetesTypes';
+import { DeploymentCondition, KubernetesJSON, KubernetesObject, KubernetesObjectKinds, ObjectMeta, ResultMetadata } from '../../types/kubernetes/kubernetesTypes';
 import { DependsOn, KubeConfig, Kustomize, NamespacedObjectKindReference } from './kustomize';
 
 /**
